Extract code actions functionality to a service
