+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/CodeActionEndpointTest.cs
@@ -458,7 +458,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         var documentPath = new Uri("C:/path/to/Page.razor");
         var codeDocument = CreateCodeDocument("@code {}");
         var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var codeActionEndpoint = CreateEndpoint(razorCodeActionProviders: [CreateRazorCodeActionProvider()]);
+        var codeActionService = CreateService(razorCodeActionProviders: [CreateRazorCodeActionProvider()]);
 
         var initialRange = VsLspFactory.CreateZeroWidthRange(0, 1);
         var selectionRange = VsLspFactory.CreateZeroWidthRange(0, 5);
@@ -473,7 +473,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         };
 
         // Act
-        var razorCodeActionContext = await codeActionEndpoint.GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, DisposalToken);
+        var razorCodeActionContext = await codeActionService.GetTestAccessor().GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, supportsCodeActionResolve: false, DisposalToken);
 
         // Assert
         Assert.NotNull(razorCodeActionContext);
@@ -487,7 +487,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         var documentPath = new Uri("C:/path/to/Page.razor");
         var codeDocument = CreateCodeDocument("@code {}");
         var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var codeActionEndpoint = CreateEndpoint(razorCodeActionProviders: [CreateRazorCodeActionProvider()]);
+        var codeActionService = CreateService(razorCodeActionProviders: [CreateRazorCodeActionProvider()]);
 
         var initialRange = VsLspFactory.CreateZeroWidthRange(0, 1);
         var request = new VSCodeActionParams()
@@ -501,7 +501,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         };
 
         // Act
-        var razorCodeActionContext = await codeActionEndpoint.GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, DisposalToken);
+        var razorCodeActionContext = await codeActionService.GetTestAccessor().GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, supportsCodeActionResolve: false, DisposalToken);
 
         // Assert
         Assert.NotNull(razorCodeActionContext);
@@ -516,7 +516,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         var codeDocument = CreateCodeDocument("@code {}");
         var documentContext = CreateDocumentContext(documentPath, codeDocument);
 
-        var codeActionEndpoint = CreateEndpoint(csharpCodeActionProviders: [CreateCSharpCodeActionProvider()]);
+        var codeActionService = CreateService(csharpCodeActionProviders: [CreateCSharpCodeActionProvider()]);
 
         var initialRange = VsLspFactory.CreateZeroWidthRange(0, 1);
         var request = new VSCodeActionParams()
@@ -526,11 +526,11 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
             Context = new VSInternalCodeActionContext()
         };
 
-        var context = await codeActionEndpoint.GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, DisposalToken);
+        var context = await codeActionService.GetTestAccessor().GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, supportsCodeActionResolve: false, DisposalToken);
         Assert.NotNull(context);
 
         // Act
-        var results = await codeActionEndpoint.GetCodeActionsFromLanguageServerAsync(RazorLanguageKind.CSharp, context, Guid.Empty, cancellationToken: default);
+        var results = await codeActionService.GetTestAccessor().GetCodeActionsFromLanguageServerAsync(RazorLanguageKind.CSharp, context, Guid.Empty, cancellationToken: default);
 
         // Assert
         Assert.Empty(results);
@@ -546,7 +546,7 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         var documentContext = CreateDocumentContext(documentPath, codeDocument);
         var projectedRange = VsLspFactory.CreateZeroWidthRange(15, 2);
 
-        var codeActionEndpoint = CreateEndpoint(
+        var codeActionService = CreateService(
             documentMappingService: CreateDocumentMappingService(projectedRange.ToLinePositionSpan()),
             csharpCodeActionProviders: [CreateCSharpCodeActionProvider()],
             clientConnection: TestClientConnection.Instance);
@@ -562,11 +562,11 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
             }
         };
 
-        var context = await codeActionEndpoint.GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, DisposalToken);
+        var context = await codeActionService.GetTestAccessor().GenerateRazorCodeActionContextAsync(request, documentContext.Snapshot, supportsCodeActionResolve: false, DisposalToken);
         Assert.NotNull(context);
 
         // Act
-        var results = await codeActionEndpoint.GetCodeActionsFromLanguageServerAsync(RazorLanguageKind.CSharp, context, Guid.Empty, cancellationToken: DisposalToken);
+        var results = await codeActionService.GetTestAccessor().GetCodeActionsFromLanguageServerAsync(RazorLanguageKind.CSharp, context, Guid.Empty, cancellationToken: DisposalToken);
 
         // Assert
         var result = Assert.Single(results);
@@ -593,22 +593,42 @@ public class CodeActionEndpointTest(ITestOutputHelper testOutput) : LanguageServ
         LanguageServerFeatureOptions? languageServerFeatureOptions = null,
         bool supportsCodeActionResolve = false)
     {
+        var codeActionsService = CreateService(
+            documentMappingService,
+            razorCodeActionProviders,
+            csharpCodeActionProviders,
+            htmlCodeActionProviders,
+            clientConnection,
+            languageServerFeatureOptions);
+
+        return new CodeActionEndpoint(
+            codeActionsService,
+            NoOpTelemetryReporter.Instance)
+        {
+            _supportsCodeActionResolve = supportsCodeActionResolve
+        };
+    }
+
+    private CodeActionsService CreateService(
+       IDocumentMappingService? documentMappingService = null,
+        ImmutableArray<IRazorCodeActionProvider> razorCodeActionProviders = default,
+        ImmutableArray<ICSharpCodeActionProvider> csharpCodeActionProviders = default,
+        ImmutableArray<IHtmlCodeActionProvider> htmlCodeActionProviders = default,
+        IClientConnection? clientConnection = null,
+        LanguageServerFeatureOptions? languageServerFeatureOptions = null)
+    {
         var delegatedCodeActionsProvider = new DelegatedCodeActionsProvider(
             clientConnection ?? StrictMock.Of<IClientConnection>(),
             NoOpTelemetryReporter.Instance,
             LoggerFactory);
 
-        return new CodeActionEndpoint(
+        return new CodeActionsService(
             documentMappingService ?? CreateDocumentMappingService(),
             razorCodeActionProviders.NullToEmpty(),
             csharpCodeActionProviders.NullToEmpty(),
             htmlCodeActionProviders.NullToEmpty(),
             delegatedCodeActionsProvider,
-            languageServerFeatureOptions ?? StrictMock.Of<LanguageServerFeatureOptions>(x => x.SupportsFileManipulation == true),
-            NoOpTelemetryReporter.Instance)
-        {
-            _supportsCodeActionResolve = supportsCodeActionResolve
-        };
+            languageServerFeatureOptions ?? StrictMock.Of<LanguageServerFeatureOptions>(x => x.SupportsFileManipulation == true));
     }
 
     private static IDocumentMappingService CreateDocumentMappingService(LinePositionSpan? projectedRange = null)
